// src/shared/types/fullTasks.types.ts
export interface IFullTasks {
	_id?: string
	who?: string
	task1?: string
	task2?: string
	task3?: string
	task4?: string
	task5?: string
	task6?: string
	task7?: string
	task8?: string
	task9?: string
	task10?: string
	task11?: string
	task12?: string
	task13?: string
	task14?: string
	task15?: string
	task16?: string
	task17?: string
	task18?: string
	task19?: string
	task19_2?: string
	task30?: string
	task31?: string
	task32?: string
	task33?: string
	task34?: string
	task35?: string
	task36?: string
	task37?: string
	task38?: string
	task39?: string
	task40?: string
	task41?: string
	task42?: string
	[key: `task${number}`]: string | undefined
}